+++ src/shared/state/ElementTab.cpp
@@ -18,7 +18,7 @@ ElementTab::ElementTab(size_t w, size_t h,std::vector<Element*> elementList){
 }
 
 ElementTab::~ElementTab(){
-    for(int i=0 ;i<elementList.size();i++){
+    for(size_t i=0 ;i<elementList.size();i++){
         delete elementList[i];
         elementList[i]=0;
         
@@ -42,7 +42,7 @@ void ElementTab::addElement(Element* elmt){
 }
 
 Element* const ElementTab::getLocatedElement(Position *loc){
-    for(int i=0;i<elementList.size();i++){
+    for(size_t i=0;i<elementList.size();i++){
         Position actu_pos=elementList[i]->getPosition();
         if(actu_pos.getX()==loc->getX() && actu_pos.getY()==loc->getY()){
             return(elementList[i]);
@@ -59,7 +59,7 @@ void ElementTab::resize(size_t w, size_t h){
 }
 
 void ElementTab::setElement(Position* pos,Element* elmt){
-    for(int i=0;i<elementList.size();i++){
+    for(size_t i=0;i<elementList.size();i++){
         if (elementList[i]==elmt){
             elementList[i]->setPosition(*pos);
         }
