+++ src/shared/state/Position.h
@@ -2,6 +2,7 @@
 #ifndef STATE__POSITION__H
 #define STATE__POSITION__H
 
+#include <stdlib.h>
 
 namespace state {
 
@@ -9,15 +10,15 @@ namespace state {
   class Position {
     // Attributes
   private:
-    int x;
-    int y;
+    size_t x;
+    size_t y;
     // Operations
   public:
-    Position (int x = 0, int y = 0);
-    void setX (int x);
-    void setY (int y);
-    int const getX ();
-    int const getY ();
+    Position (size_t x = 0, size_t y = 0);
+    void setX (size_t x);
+    void setY (size_t y);
+    size_t const getX ();
+    size_t const getY ();
     // Setters and Getters
   };
